requirements/base.txt
# @TODO: install --no-deps version of arcgis and only install needed dependencies
arcgis==2.0.0
crispy-tailwind==0.5.0
Django==4.0.6
django-crispy-forms==1.14.0
django-environ==0.9.0
django-multiupload==0.6.1
django-tailwind==3.4.0
docusign-esign==3.17.0
exif==1.3.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
